Changeset 328 @ 2014.06.19

+++ aCis_gameserver/java/net/sf/l2j/gameserver/network/clientpackets/SetPrivateStoreListBuy.java
@@ -42,7 +42,7 @@ public final class SetPrivateStoreListBuy extends L2GameClientPacket
 			int itemId = readD();
 			readH(); // TODO analyse this
 			readH(); // TODO analyse this
-			long cnt = readD();
+			int cnt = readD();
 			int price = readD();
 			
 			if (itemId < 1 || cnt < 1 || price < 0)
@@ -50,7 +50,7 @@ public final class SetPrivateStoreListBuy extends L2GameClientPacket
 				_items = null;
 				return;
 			}
-			_items[i] = new Item(itemId, (int) cnt, price);
+			_items[i] = new Item(itemId, cnt, price);
 		}
 	}
